pyproject.toml
[tool.poetry]
name = "search-web-tool"
version = "0.1.0"
description = ""
authors = ["Gowtham Satti <sattigowtham1999@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
naptha-sdk = {git = "https://github.com/NapthaAI/naptha-sdk.git"}
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
